Moving all code into modules and simplifying
