+++ lib/handlers/index.js
@@ -1,0 +1,2 @@
+exports.subscriptions = require('./subscriptions');
+exports.notifications = require('./notifications');
