+++ lib/index.js
@@ -1,0 +1,2 @@
+exports.handlers = require('./handlers');
+exports.server = require('./server');
