+++ package.json
@@ -2,8 +2,9 @@
   "name": "push-messaging-and-notifications",
   "version": "1.0.0",
   "description": "Push Messaging and Notification Sample ===",
-  "main": "index.js",
+  "main": "lib/index.js",
   "scripts": {
+    "start": "node index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "Sean O'Hara",
@@ -12,6 +13,7 @@
     "body-parser": "^1.15.0",
     "express": "^4.13.4",
     "express-handlebars": "^3.0.0",
+    "lodash": "^4.6.1",
     "mongoose": "^4.4.5",
     "request": "^2.69.0"
   }
